requirements.txt
astroid==2.0.4
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
Flask==1.0.2
google-api-core==1.4.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-storage==1.12.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pylint==2.1.1
pytz==2018.5
requests==2.19.1
rsa==3.4.2
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
